+++ test/foundry/Raffle.fulfillRandomWords.t.sol
@@ -35,9 +35,9 @@ contract Raffle_FulfillRandomWords_Test is TestHelpers {
         vm.startPrank(user1);
         looksRareRaffle.createRaffle({
             cutoffTime: uint40(block.timestamp + 86_400),
-            minimumEntries: uint80(107),
-            maximumEntries: uint80(512),
-            maximumEntriesPerParticipant: uint80(100),
+            minimumEntries: uint64(107),
+            maximumEntries: uint64(512),
+            maximumEntriesPerParticipant: uint64(100),
             prizesTotalValue: 1 ether,
             minimumProfitBp: uint16(500),
             feeTokenAddress: address(0),
@@ -307,7 +307,7 @@ contract Raffle_FulfillRandomWords_Test is TestHelpers {
     function testFuzz_fulfillRandomWords(uint248 seed) public {
         IRaffle.PricingOption[5] memory pricingOptions = _generateStandardPricings();
         uint256 userIndex;
-        uint80 currentEntryIndex;
+        uint64 currentEntryIndex;
         while (currentEntryIndex < 107) {
             address participant = address(uint160(userIndex + 1));
             vm.deal(participant, 1 ether);
@@ -366,7 +366,7 @@ contract Raffle_FulfillRandomWords_Test is TestHelpers {
             assertFalse(winners[i].claimed);
             assertNotEq(winners[i].participant, address(0));
 
-            uint80 entryIndex = winners[i].entryIndex;
+            uint64 entryIndex = winners[i].entryIndex;
             assertFalse(winningEntries[entryIndex]);
             winningEntries[entryIndex] = true;
